docs: update GitHub repository URL
We already renamed the NPM package name from ngQuickDate to
@egston/ng-quick-date.angularjs.

Next step taken was renaming of the GitHub repository to more suitable
Egston/egston-ng-quick-date.angularjs.git.

This commit updates cloning instructions in README.md to reflect this
change.

Part of issue egston#2014: Polish NPM package ngQuickDate so that it can
be published on GitHub and do so